refactor: rename shared R2 storage helper
